# swagger.yaml
openapi: 3.0.3
info:
  title: Platform API
  version: 1.0.0
  description: API for user registration, product and blog management, resource tracking, and forums.
components:
  schemas:
    Role:
      type: string
      enum: [USER, EMPLOYEE]
    Language:
      type: string
      enum: [es-ES, en-US]
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        nickName:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        image:
          type: string
        isBloqued:
          type: boolean
        creationDate:
          type: string
          format: date-time
    ProductLanContent:
      type: object
      properties:
        lan:
          $ref: "#/components/schemas/Language"
        name:
          type: string
        description:
          type: string
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        image:
          type: string
        creationDate:
          type: string
          format: date-time
        regions:
          type: array
          items:
            type: string
            description: ISO 3166-2:ES region code
        productLanContents:
          type: array
          items:
            $ref: "#/components/schemas/ProductLanContent"
    BlogComment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        blogId:
          type: string
          format: uuid
        user:
          $ref: "#/components/schemas/User"
        comment:
          type: string
        image:
          type: string
          nullable: true
        nLikes:
          type: integer
        creationDate:
          type: string
          format: date-time
    Blog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        product:
          $ref: "#/components/schemas/Product"
        lan:
          $ref: "#/components/schemas/Language"
        title:
          type: string
        description:
          type: string
        image:
          type: string
        comments:
          type: array
          items:
            $ref: "#/components/schemas/BlogComment"
        creationDate:
          type: string
          format: date-time
    ProductRequest:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        user:
          $ref: "#/components/schemas/User"
        image:
          type: string
        description:
          type: string
        creationDate:
          type: string
          format: date-time
    BlogRequest:
      type: object
      properties:
        id:
          type: string
        user:
          $ref: "#/components/schemas/User"
        title:
          type: string
        description:
          type: string
        product:
          $ref: "#/components/schemas/Product"
        image:
          type: string
        creationDate:
          type: string
          format: date-time
...
paths:
  /users/auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                nickName:
                  type: string
                role:
                  $ref: "#/components/schemas/Role"
                password:
                  type: string
                image:
                  type: string
      tags:
        - users
      responses:
        "201":
          description: User created successfully
        "400":
          description: Email or nickname exists or invalid input
        "401":
          description: Unauthorized to create EMPLOYEE role
  /users/auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      tags:
        - users
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  bearer:
                    type: string
        "401":
          description: Invalid email or password
  /users/{id}/block:
    put:
      summary: Block a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      tags:
        - users
      responses:
        "201":
          description: User blocked
        "401":
          description: Unauthorized
  /users/{id}/unblock:
    put:
      summary: Unblock a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      tags:
        - users
      responses:
        "201":
          description: User unblocked
        "401":
          description: Unauthorized
  /users/me:
    get:
      summary: Get current user info
      tags:
        - users
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /users:
    get:
      summary: Get list of users
      parameters:
        - in: query
          name: search
          schema:
            type: string
          required: false
      tags:
        - users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /products:
    post:
      summary: Create base product structure
      tags:
        - products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                image:
                  type: string
                regions:
                  type: string
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
        "400":
          description: Bad request
        "401":
          description: Forbidden
    get:
      summary: Search products
      parameters:
        - in: query
          name: name
          schema:
            type: string
        - in: query
          name: region
          schema:
            type: string
      tags:
        - products
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "404":
          description: Not found
...
  /blogs:
    get:
      summary: Search blogs
      parameters:
        - in: query
          name: name
          schema:
            type: string
        - in: query
          name: productId
          schema:
            type: string
      tags:
        - blogs
      responses:
        "200":
          description: List of blogs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Blog"
  /requests/products:
    post:
      summary: Submit product request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                userId:
                  type: string
                image:
                  type: string
                description:
                  type: string
      tags:
        - requests
      responses:
        "201":
          description: Product request created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
    get:
      summary: Get product requests
      tags:
        - products
      responses:
        "200":
          description: List of product requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductRequest"
  /requests/blogs:
    post:
      summary: Submit blog request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                productId:
                  type: string
                userId:
                  type: string
                image:
                  type: string
                description:
                  type: string
      tags:
        - requests
      responses:
        "201":
          description: Blog request created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
    get:
      summary: Get blog requests
      tags:
        - requests
      responses:
        "200":
          description: List of blog requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogRequest"
  /requests/blogs/{id}/comment:
    post:
      summary: Add comment to blog request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                blogId:
                  type: string
                userId:
                  type: string
                comment:
                  type: string
                image:
                  type: string
                  nullable: true
      tags:
        - requests
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
...